+++ SwiftUI_Pull_to_Refresh/Extensions/UIScrollView+Extensions.swift
@@ -1,0 +1,21 @@
+//
+//  UIScrollView+Extensions.swift
+//  SwiftUI_Pull_to_Refresh
+//
+//  Created by Geri Borbás on 23/03/2022.
+//
+
+import Foundation
+import UIKit
+
+
+extension UIScrollView {
+	
+	func set(decelerationRate: CGFloat) {
+		self.decelerationRate = .init(rawValue: decelerationRate)
+	}
+	
+	func set(pagingFriction: CGFloat) {
+		self.setValue(pagingFriction, forKey: "pagingFriction")
+	}
+}